Moved Week to model package
